# services/api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: pairexx API Gateway
  description: API Gateway for pairexx microservices
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local Gateway
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /users/anonymous:
    post:
      summary: Create anonymous session
      description: Create a new anonymous user session and get tokens
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
...